# api/openapi.yaml
openapi: 3.0.3
info:
  title: Live Streaming Tipping Platform API
  description: |
    Secure, scalable live-streaming platform where performers broadcast video, 
    viewers watch, and viewers tip virtual tokens in real time.
    
    ## Key Features
    - Real-time WebRTC streaming with SFU
    - Virtual token economy with instant tipping
    - Age verification and KYC compliance
    - Automated and human moderation
    - PCI-compliant payment processing
    
    ## Authentication
    All endpoints require JWT authentication except registration, login, and public endpoints.
    Include the JWT token in the Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - General API: 1000 requests/hour per user
    - Streaming endpoints: 100 requests/minute per user
    - Payment endpoints: 10 requests/minute per user
    - Tipping: 60 tips/minute per user
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@livepanty.com
  license:
    name: Proprietary
servers:
  - url: https://api.livepanty.com/v1
    description: Production server
  - url: https://staging-api.livepanty.com/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /streams:
    get:
      tags:
        - Streaming
      summary: Get live streams
      description: Retrieve list of currently live streams
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Live streams retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamList'

    post:
      tags:
        - Streaming
      summary: Create new stream
      description: Create a new streaming session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStreamRequest'
      responses:
        '201':
          description: Stream created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamSession'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get all users (Admin only)
      description: Retrieve list of all users with pagination
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [viewer, performer, admin, moderator]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, banned, pending_verification]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserSummary'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        country:
          type: string
        email_verified:
          type: boolean
        last_login_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        token_balance:
          type: integer
        kyc_status:
          type: string

    # ============================================================================
    # KYC SCHEMAS
    # ============================================================================
...
    CreateStreamRequest:
      type: object
      required:
        - title
        - category
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: "My Amazing Stream"
        description:
          type: string
          maxLength: 1000
        category:
          type: string
          example: "entertainment"
        tags:
          type: array
          items:
            type: string
          maxItems: 10
        is_private:
          type: boolean
          default: false
        is_age_restricted:
          type: boolean
          default: true
        tip_enabled:
          type: boolean
          default: true
        chat_enabled:
          type: boolean
          default: true
        recording_enabled:
          type: boolean
          default: false
    StreamSession:
      type: object
      properties:
        stream_id:
          type: string
          format: uuid
        signaling_url:
          type: string
          format: uri
        sfu_room_id:
          type: string
        turn_servers:
          type: array
          items:
            $ref: '#/components/schemas/TURNServer'
        viewer_token:
          type: string
        expires_at:
          type: string
          format: date-time
    TURNServer:
      type: object
      properties:
        urls:
          type: array
          items:
            type: string
        username:
          type: string
        credential:
          type: string
...
    StreamList:
      type: object
      properties:
        streams:
          type: array
          items:
            $ref: '#/components/schemas/StreamSummary'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    StreamSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        host_id:
          type: string
          format: uuid
        host_name:
          type: string
        title:
          type: string
        category:
          type: string
        status:
          type: string
        is_private:
          type: boolean
        viewer_count:
          type: integer
        total_tokens_received:
          type: integer
        thumbnail_url:
          type: string
          format: uri
        started_at:
          type: string
          format: date-time
...